fix(#63): isCancellableByTherapist dans module pur (plus de supabase dans le bundle client)
Déplace isCancellableByTherapist de appointment-eligibility.ts vers utils/date.ts
(module pur sans I/O). AppointmentCard (island client:load) l'importait transitivement
via appointment-eligibility → manual-slots → supabaseAdmin, ce qui bundlait supabase
côté client et faisait échouer l'hydration ('supabaseUrl is required').

appointment-eligibility.ts re-exporte la fonction pour les importations serveur
([id].ts) inchangées. Vérifié : 0 occurrence de createClient/supabase dans le chunk
client AppointmentCard.
